+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/v2021_08_01/aio/_configuration.py
@@ -44,7 +44,7 @@ class ContainerServiceClientConfiguration(Configuration):
 
         self.credential = credential
         self.subscription_id = subscription_id
-        self.api_version = "2021-07-01"
+        self.api_version = "2021-08-01"
         self.credential_scopes = kwargs.pop('credential_scopes', ['https://management.azure.com/.default'])
         kwargs.setdefault('sdk_moniker', 'mgmt-containerservice/{}'.format(VERSION))
         self._configure(**kwargs)