+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/v2021_08_01/aio/_container_service_client.py
@@ -6,17 +6,15 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from typing import TYPE_CHECKING
+from typing import Any, Optional, TYPE_CHECKING
 
-from azure.mgmt.core import ARMPipelineClient
+from azure.core.pipeline.transport import AsyncHttpResponse, HttpRequest
+from azure.mgmt.core import AsyncARMPipelineClient
 from msrest import Deserializer, Serializer
 
 if TYPE_CHECKING:
     # pylint: disable=unused-import,ungrouped-imports
-    from typing import Any, Optional
-
-    from azure.core.credentials import TokenCredential
-    from azure.core.pipeline.transport import HttpRequest, HttpResponse
+    from azure.core.credentials_async import AsyncTokenCredential
 
 from ._configuration import ContainerServiceClientConfiguration
 from .operations import Operations
@@ -26,28 +24,31 @@ from .operations import AgentPoolsOperations
 from .operations import PrivateEndpointConnectionsOperations
 from .operations import PrivateLinkResourcesOperations
 from .operations import ResolvePrivateLinkServiceIdOperations
-from . import models
+from .operations import SnapshotsOperations
+from .. import models
 
 
 class ContainerServiceClient(object):
     """The Container Service Client.
 
     :ivar operations: Operations operations
-    :vartype operations: azure.mgmt.containerservice.v2021_07_01.operations.Operations
+    :vartype operations: azure.mgmt.containerservice.v2021_08_01.aio.operations.Operations
     :ivar managed_clusters: ManagedClustersOperations operations
-    :vartype managed_clusters: azure.mgmt.containerservice.v2021_07_01.operations.ManagedClustersOperations
+    :vartype managed_clusters: azure.mgmt.containerservice.v2021_08_01.aio.operations.ManagedClustersOperations
     :ivar maintenance_configurations: MaintenanceConfigurationsOperations operations
-    :vartype maintenance_configurations: azure.mgmt.containerservice.v2021_07_01.operations.MaintenanceConfigurationsOperations
+    :vartype maintenance_configurations: azure.mgmt.containerservice.v2021_08_01.aio.operations.MaintenanceConfigurationsOperations
     :ivar agent_pools: AgentPoolsOperations operations
-    :vartype agent_pools: azure.mgmt.containerservice.v2021_07_01.operations.AgentPoolsOperations
+    :vartype agent_pools: azure.mgmt.containerservice.v2021_08_01.aio.operations.AgentPoolsOperations
     :ivar private_endpoint_connections: PrivateEndpointConnectionsOperations operations
-    :vartype private_endpoint_connections: azure.mgmt.containerservice.v2021_07_01.operations.PrivateEndpointConnectionsOperations
+    :vartype private_endpoint_connections: azure.mgmt.containerservice.v2021_08_01.aio.operations.PrivateEndpointConnectionsOperations
     :ivar private_link_resources: PrivateLinkResourcesOperations operations
-    :vartype private_link_resources: azure.mgmt.containerservice.v2021_07_01.operations.PrivateLinkResourcesOperations
+    :vartype private_link_resources: azure.mgmt.containerservice.v2021_08_01.aio.operations.PrivateLinkResourcesOperations
     :ivar resolve_private_link_service_id: ResolvePrivateLinkServiceIdOperations operations
-    :vartype resolve_private_link_service_id: azure.mgmt.containerservice.v2021_07_01.operations.ResolvePrivateLinkServiceIdOperations
+    :vartype resolve_private_link_service_id: azure.mgmt.containerservice.v2021_08_01.aio.operations.ResolvePrivateLinkServiceIdOperations
+    :ivar snapshots: SnapshotsOperations operations
+    :vartype snapshots: azure.mgmt.containerservice.v2021_08_01.aio.operations.SnapshotsOperations
     :param credential: Credential needed for the client to connect to Azure.
-    :type credential: ~azure.core.credentials.TokenCredential
+    :type credential: ~azure.core.credentials_async.AsyncTokenCredential
     :param subscription_id: Subscription credentials which uniquely identify Microsoft Azure subscription. The subscription ID forms part of the URI for every service call.
     :type subscription_id: str
     :param str base_url: Service URL
@@ -56,16 +57,15 @@ class ContainerServiceClient(object):
 
     def __init__(
         self,
-        credential,  # type: "TokenCredential"
-        subscription_id,  # type: str
-        base_url=None,  # type: Optional[str]
-        **kwargs  # type: Any
-    ):
-        # type: (...) -> None
+        credential: "AsyncTokenCredential",
+        subscription_id: str,
+        base_url: Optional[str] = None,
+        **kwargs: Any
+    ) -> None:
         if not base_url:
             base_url = 'https://management.azure.com'
         self._config = ContainerServiceClientConfiguration(credential, subscription_id, **kwargs)
-        self._client = ARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
+        self._client = AsyncARMPipelineClient(base_url=base_url, config=self._config, **kwargs)
 
         client_models = {k: v for k, v in models.__dict__.items() if isinstance(v, type)}
         self._serialize = Serializer(client_models)
@@ -86,34 +86,32 @@ class ContainerServiceClient(object):
             self._client, self._config, self._serialize, self._deserialize)
         self.resolve_private_link_service_id = ResolvePrivateLinkServiceIdOperations(
             self._client, self._config, self._serialize, self._deserialize)
+        self.snapshots = SnapshotsOperations(
+            self._client, self._config, self._serialize, self._deserialize)
 
-    def _send_request(self, http_request, **kwargs):
-        # type: (HttpRequest, Any) -> HttpResponse
+    async def _send_request(self, http_request: HttpRequest, **kwargs: Any) -> AsyncHttpResponse:
         """Runs the network request through the client's chained policies.
 
         :param http_request: The network request you want to make. Required.
         :type http_request: ~azure.core.pipeline.transport.HttpRequest
         :keyword bool stream: Whether the response payload will be streamed. Defaults to True.
         :return: The response of your network call. Does not do error handling on your response.
-        :rtype: ~azure.core.pipeline.transport.HttpResponse
+        :rtype: ~azure.core.pipeline.transport.AsyncHttpResponse
         """
         path_format_arguments = {
             'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
         }
         http_request.url = self._client.format_url(http_request.url, **path_format_arguments)
         stream = kwargs.pop("stream", True)
-        pipeline_response = self._client._pipeline.run(http_request, stream=stream, **kwargs)
+        pipeline_response = await self._client._pipeline.run(http_request, stream=stream, **kwargs)
         return pipeline_response.http_response
 
-    def close(self):
-        # type: () -> None
-        self._client.close()
+    async def close(self) -> None:
+        await self._client.close()
 
-    def __enter__(self):
-        # type: () -> ContainerServiceClient
-        self._client.__enter__()
+    async def __aenter__(self) -> "ContainerServiceClient":
+        await self._client.__aenter__()
         return self
 
-    def __exit__(self, *exc_details):
-        # type: (Any) -> None
-        self._client.__exit__(*exc_details)
+    async def __aexit__(self, *exc_details) -> None:
+        await self._client.__aexit__(*exc_details)