+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/v2021_08_01/aio/operations/_agent_pools_operations.py
@@ -28,7 +28,7 @@ class AgentPoolsOperations:
     instantiates it for you and attaches it as an attribute.
 
     :ivar models: Alias to model classes used in this operation group.
-    :type models: ~azure.mgmt.containerservice.v2021_07_01.models
+    :type models: ~azure.mgmt.containerservice.v2021_08_01.models
     :param client: Client for service requests.
     :param config: Configuration of service client.
     :param serializer: An object model serializer.
@@ -59,7 +59,7 @@ class AgentPoolsOperations:
         :type resource_name: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: An iterator like instance of either AgentPoolListResult or the result of cls(response)
-        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.mgmt.containerservice.v2021_07_01.models.AgentPoolListResult]
+        :rtype: ~azure.core.async_paging.AsyncItemPaged[~azure.mgmt.containerservice.v2021_08_01.models.AgentPoolListResult]
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["_models.AgentPoolListResult"]
@@ -67,7 +67,7 @@ class AgentPoolsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2021-07-01"
+        api_version = "2021-08-01"
         accept = "application/json"
 
         def prepare_request(next_link=None):
@@ -138,7 +138,7 @@ class AgentPoolsOperations:
         :type agent_pool_name: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: AgentPool, or the result of cls(response)
-        :rtype: ~azure.mgmt.containerservice.v2021_07_01.models.AgentPool
+        :rtype: ~azure.mgmt.containerservice.v2021_08_01.models.AgentPool
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["_models.AgentPool"]
@@ -146,7 +146,7 @@ class AgentPoolsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2021-07-01"
+        api_version = "2021-08-01"
         accept = "application/json"
 
         # Construct URL
@@ -196,7 +196,7 @@ class AgentPoolsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2021-07-01"
+        api_version = "2021-08-01"
         content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
@@ -261,7 +261,7 @@ class AgentPoolsOperations:
         :param agent_pool_name: The name of the agent pool.
         :type agent_pool_name: str
         :param parameters: The agent pool to create or update.
-        :type parameters: ~azure.mgmt.containerservice.v2021_07_01.models.AgentPool
+        :type parameters: ~azure.mgmt.containerservice.v2021_08_01.models.AgentPool
         :keyword callable cls: A custom type or function that will be passed the direct response
         :keyword str continuation_token: A continuation token to restart a poller from a saved state.
         :keyword polling: By default, your polling method will be AsyncARMPolling.
@@ -269,7 +269,7 @@ class AgentPoolsOperations:
         :paramtype polling: bool or ~azure.core.polling.AsyncPollingMethod
         :keyword int polling_interval: Default waiting time between two polls for LRO operations if no Retry-After header is present.
         :return: An instance of AsyncLROPoller that returns either AgentPool or the result of cls(response)
-        :rtype: ~azure.core.polling.AsyncLROPoller[~azure.mgmt.containerservice.v2021_07_01.models.AgentPool]
+        :rtype: ~azure.core.polling.AsyncLROPoller[~azure.mgmt.containerservice.v2021_08_01.models.AgentPool]
         :raises ~azure.core.exceptions.HttpResponseError:
         """
         polling = kwargs.pop('polling', True)  # type: Union[bool, AsyncPollingMethod]
@@ -332,7 +332,7 @@ class AgentPoolsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2021-07-01"
+        api_version = "2021-08-01"
         accept = "application/json"
 
         # Construct URL
@@ -456,7 +456,7 @@ class AgentPoolsOperations:
         :type agent_pool_name: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: AgentPoolUpgradeProfile, or the result of cls(response)
-        :rtype: ~azure.mgmt.containerservice.v2021_07_01.models.AgentPoolUpgradeProfile
+        :rtype: ~azure.mgmt.containerservice.v2021_08_01.models.AgentPoolUpgradeProfile
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["_models.AgentPoolUpgradeProfile"]
@@ -464,7 +464,7 @@ class AgentPoolsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2021-07-01"
+        api_version = "2021-08-01"
         accept = "application/json"
 
         # Construct URL
@@ -519,7 +519,7 @@ class AgentPoolsOperations:
         :type resource_name: str
         :keyword callable cls: A custom type or function that will be passed the direct response
         :return: AgentPoolAvailableVersions, or the result of cls(response)
-        :rtype: ~azure.mgmt.containerservice.v2021_07_01.models.AgentPoolAvailableVersions
+        :rtype: ~azure.mgmt.containerservice.v2021_08_01.models.AgentPoolAvailableVersions
         :raises: ~azure.core.exceptions.HttpResponseError
         """
         cls = kwargs.pop('cls', None)  # type: ClsType["_models.AgentPoolAvailableVersions"]
@@ -527,7 +527,7 @@ class AgentPoolsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2021-07-01"
+        api_version = "2021-08-01"
         accept = "application/json"
 
         # Construct URL
@@ -575,7 +575,7 @@ class AgentPoolsOperations:
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2021-07-01"
+        api_version = "2021-08-01"
         accept = "application/json"
 
         # Construct URL