+++ src/aks-preview/azext_aks_preview/vendored_sdks/azure_mgmt_preview_aks/v2021_08_01/operations/_resolve_private_link_service_id_operations.py
@@ -22,14 +22,14 @@ if TYPE_CHECKING:
     T = TypeVar('T')
     ClsType = Optional[Callable[[PipelineResponse[HttpRequest, HttpResponse], T, Dict[str, Any]], Any]]
 
-class PrivateLinkResourcesOperations(object):
-    """PrivateLinkResourcesOperations operations.
+class ResolvePrivateLinkServiceIdOperations(object):
+    """ResolvePrivateLinkServiceIdOperations operations.
 
     You should not instantiate this class directly. Instead, you should create a Client instance that
     instantiates it for you and attaches it as an attribute.
 
     :ivar models: Alias to model classes used in this operation group.
-    :type models: ~azure.mgmt.containerservice.v2021_07_01.models
+    :type models: ~azure.mgmt.containerservice.v2021_08_01.models
     :param client: Client for service requests.
     :param config: Configuration of service client.
     :param serializer: An object model serializer.
@@ -44,37 +44,40 @@ class PrivateLinkResourcesOperations(object):
         self._deserialize = deserializer
         self._config = config
 
-    def list(
+    def post(
         self,
         resource_group_name,  # type: str
         resource_name,  # type: str
+        parameters,  # type: "_models.PrivateLinkResource"
         **kwargs  # type: Any
     ):
-        # type: (...) -> "_models.PrivateLinkResourcesListResult"
-        """Gets a list of private link resources in the specified managed cluster.
+        # type: (...) -> "_models.PrivateLinkResource"
+        """Gets the private link service ID for the specified managed cluster.
 
-        To learn more about private clusters, see:
-        https://docs.microsoft.com/azure/aks/private-clusters.
+        Gets the private link service ID for the specified managed cluster.
 
         :param resource_group_name: The name of the resource group.
         :type resource_group_name: str
         :param resource_name: The name of the managed cluster resource.
         :type resource_name: str
+        :param parameters: Parameters required in order to resolve a private link service ID.
+        :type parameters: ~azure.mgmt.containerservice.v2021_08_01.models.PrivateLinkResource
         :keyword callable cls: A custom type or function that will be passed the direct response
-        :return: PrivateLinkResourcesListResult, or the result of cls(response)
-        :rtype: ~azure.mgmt.containerservice.v2021_07_01.models.PrivateLinkResourcesListResult
+        :return: PrivateLinkResource, or the result of cls(response)
+        :rtype: ~azure.mgmt.containerservice.v2021_08_01.models.PrivateLinkResource
         :raises: ~azure.core.exceptions.HttpResponseError
         """
-        cls = kwargs.pop('cls', None)  # type: ClsType["_models.PrivateLinkResourcesListResult"]
+        cls = kwargs.pop('cls', None)  # type: ClsType["_models.PrivateLinkResource"]
         error_map = {
             401: ClientAuthenticationError, 404: ResourceNotFoundError, 409: ResourceExistsError
         }
         error_map.update(kwargs.pop('error_map', {}))
-        api_version = "2021-07-01"
+        api_version = "2021-08-01"
+        content_type = kwargs.pop("content_type", "application/json")
         accept = "application/json"
 
         # Construct URL
-        url = self.list.metadata['url']  # type: ignore
+        url = self.post.metadata['url']  # type: ignore
         path_format_arguments = {
             'subscriptionId': self._serialize.url("self._config.subscription_id", self._config.subscription_id, 'str'),
             'resourceGroupName': self._serialize.url("resource_group_name", resource_group_name, 'str', min_length=1),
@@ -88,9 +91,13 @@ class PrivateLinkResourcesOperations(object):
 
         # Construct headers
         header_parameters = {}  # type: Dict[str, Any]
+        header_parameters['Content-Type'] = self._serialize.header("content_type", content_type, 'str')
         header_parameters['Accept'] = self._serialize.header("accept", accept, 'str')
 
-        request = self._client.get(url, query_parameters, header_parameters)
+        body_content_kwargs = {}  # type: Dict[str, Any]
+        body_content = self._serialize.body(parameters, 'PrivateLinkResource')
+        body_content_kwargs['content'] = body_content
+        request = self._client.post(url, query_parameters, header_parameters, **body_content_kwargs)
         pipeline_response = self._client._pipeline.run(request, stream=False, **kwargs)
         response = pipeline_response.http_response
 
@@ -98,10 +105,10 @@ class PrivateLinkResourcesOperations(object):
             map_error(status_code=response.status_code, response=response, error_map=error_map)
             raise HttpResponseError(response=response, error_format=ARMErrorFormat)
 
-        deserialized = self._deserialize('PrivateLinkResourcesListResult', pipeline_response)
+        deserialized = self._deserialize('PrivateLinkResource', pipeline_response)
 
         if cls:
             return cls(pipeline_response, deserialized, {})
 
         return deserialized
-    list.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ContainerService/managedClusters/{resourceName}/privateLinkResources'}  # type: ignore
+    post.metadata = {'url': '/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.ContainerService/managedClusters/{resourceName}/resolvePrivateLinkServiceId'}  # type: ignore